+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Dropped support for Python 3.7 or lower
+- Remove `--disk` CLI option, just use `--output-root -` instead
 
 ### Fixed